Simplify interface by consolidating inputs
Simplify usage of the GitHub Action by consolidating
the action's inputs and removing those that do not
have user-facing value.

Drop the openapi-yaml-directory since this is not
user-facing and is generated in a temp directory
of the GitHub Actions worker, so we can specify
a value within our script.

Combine the api-docs-directory and api-doc-filename
inputs into a single api-docs-filepath input, since
the directory can be inferred from this.

This input changes are breaking changes, so
will increment the version and generate a new
release with a changelog reflecting this.

Update README to reflect updates and add a
Technologies section with links to tech used.