Documented all prerequisites
In order for clients to be able to use this GitHub Action,
their repository needs to be supported by GitHub Pages,
and they need to enable GitHub Pages with GitHub Actions
under their repository's settings.

Expanded README with prerequisites, simplifying and
combining the subsequent requirements section that
described the two run modes of this Action.