Use caller repository Git context
Prior scripts worked when run locally but not
when run remotely on GitHub, due to the local
directory the generate-single-spec-doc.sh was
run in not being in the right directory that
has Git context on the caller repository, so
that it is unable to Git fetch the referenced
OpenAPI spec files.

Updated to use the workspace directory from
the `github.workspace` context.

Ref: https://docs.github.com/en/actions/learn-github-actions/contexts#github-context

Simplify action.yml to retrieve all scripts from
the openapi-github-pages-action repository.