Fix GitHub token integration
Prior way of passing the GitHub token required for
making Git CLI commands did not work, with the
execution throwing the error:

```
Error:
/home/runner/work/openapi-github-pages-action/openapi-github-pages-action/./action.yml
(Line: 55, Col: 19): Unrecognized named-value: 'secrets'. Located at
position 1 within expression: secrets.GITHUB_TOKEN
```

As per
https://stackoverflow.com/questions/69493352/error-passing-secret-to-reusable-workflow
and
https://docs.github.com/en/actions/security-guides/automatic-token-authentication#about-the-github_token-secret,
it looks like when referencing secrets.GITHUB_TOKEN, the caller also
needs to specify that they're passing the token to you, which is
an additional user-required config and point for error.

We can instead directly reference github.token which only needs
to be specified on the called action's side.